refactor: 💡 profile card
