--- src/entities/profile/index.ts
@@ -1,1 +1,0 @@
-export * as Profile from './model/profile.model';
