+++ src/features/view-crew-profile/index.ts
@@ -1,1 +1,4 @@
+export { default as useViewCrewProfile } from './api/use-view-crew-profile';
 export { default as ViewCrewProfile } from './ui/view-crew-profile.component';
+export type { ProfileCardType } from './model/profile.model';
+export { extractScore, getRegistrationDate } from './lib/profile.lib';
