+++ src/features/view-crew-profile/model/profile.model.ts
@@ -1,0 +1,10 @@
+export type ProfileCardType = {
+  avatarBodyUri: string;
+  avatarFaceUri: string;
+  combinePositionX?: number;
+  combinePositionY?: number;
+  nickname: string;
+  introduction: string;
+  score: number;
+  registrationDate: string;
+};
