--- src/shared/ui/components/profile-card/index.ts
@@ -1,1 +1,0 @@
-export { default as ProfileCard } from './profile-card.component';
